+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project closely adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.3.5
+- Change - remove cli option show-status to simplify code
+- Change - rename default arg on report command
+
 ## 1.3.4
 - Fix - access step.is_substep only in after_step